- Bollinger Bands %b module properties bug fixed
git-svn-id: https://chartsy.svn.sourceforge.net/svnroot/chartsy/trunk@310 7741bc21-d5e8-4718-ad21-8ba3422b8f82

+++ Chartsy/Bollinger Bands %b/src/org/chartsy/bollingerb/IndicatorNode.java
@@ -66,7 +66,7 @@ public class IndicatorNode
                     "Std. Dev. High", // property name
                     "Sets the std. dev. high period value", // property description
                     IndicatorProperties.class, // properties class
-                    int.class, // property class
+                    double.class, // property class
                     null, // property editor class (null if none)
                     "getStdHigh", // get method name
                     "setStdHigh", // set method name
@@ -77,7 +77,7 @@ public class IndicatorNode
                     "Std. Dev. Low", // property name
                     "Sets the std. dev. low period value", // property description
                     IndicatorProperties.class, // properties class
-                    int.class, // property class
+                    double.class, // property class
                     null, // property editor class (null if none)
                     "getStdLow", // get method name
                     "setStdLow", // set method name
